Build against Qt6 on F40+
In Fedora 40+ there is KDE 6, with which the Qt6 build of Strawberry
finally looks OK (with KDE 5 there is some UI ugliness). Thus, switch to
building against Qt6 on these newer Fedora versions.

There is one minor regression in the Qt6 build on KDE 6 (ampersands in
left bar titles) that is being tracked upstream as:
https://bugs.kde.org/show_bug.cgi?id=483488

Signed-off-by: Ondrej Mosnáček <omosnacek@gmail.com>